+++ snaily-cad.sql
@@ -33,7 +33,7 @@ CREATE TABLE `911calls` (
   `name` varchar(255) NOT NULL,
   `location` varchar(255) NOT NULL,
   `status` varchar(255) NOT NULL,
-  `assigned_unit` varchar(255) NOT NULL
+  `assigned_unit` text NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4;
 
 -- --------------------------------------------------------