+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ## 1.2.4
 
 - New notifications system
+- Automatically change unit status to `10-97` when assigned to a call
+- Fixed length for assigning units to a call
+- Added callsign to update 911 call "assign unit"
 
 ## 1.2.3